Add additional linter checks (#331)
* Add additional linter params

* Bump linter version

* Fix whitespace lint errors

* Format source tree

* Fix wsl linting errors, pt. 1

* Fix wsl linting errors pt. 2

* Fix wsl linting errors pt. 3

* Fix wsl linting errors pt. 4

* Fix wsl linting errors pt. 5

* Fix wsl linting errors pt. 6

* Fix wsl linting errors pt. 7

* Fix wsl linting errors pt. 8

* Fix wsl linting errors pt. 9

* Fix wsl linting errors pt. 10

* Fix wsl linting errors pt. 11

* Fix wastedassign linting errors

* Fix unconvert linting errors

* Fix misspell linting errors

* Fix gofmt linting errors

* Fix ifshort linting errors

* Remove the nilnil linter until the horrendous blockchain interface is sorted out

* Fix errname linting errors

* Fix makezero linting errors

* Fix prealloc linting errors

* Fix forcetypeassert linting errors

* Fix lll linting errors

* Fix predeclared linting errors

* Fix stylecheck linting errors

* Fix nolintlint linting errors

* Fix nlreturn linting errors

* Fix dupl linting errors

* Fix thelper linting errors

* Fix gosec linting errors

* Remove gomnd check

* Fix errorlint linting errors

* Fix leftover linting errors

* Fix incorrect conditional check

* Add lint command to makefile

* Drop appending in findPeersCall

+++ consensus/ibft/snapshot.go
@@ -28,6 +28,7 @@ func (i *Ibft) setupSnapshot() error {
 
 	header := i.blockchain.Header()
 	meta, err := i.getSnapshotMetadata()
+
 	if err != nil {
 		return err
 	}
@@ -48,12 +49,14 @@ func (i *Ibft) setupSnapshot() error {
 	currentEpoch := header.Number / i.epochSize
 	metaEpoch := meta.LastBlock / i.epochSize
 	snapshot, _ := i.getSnapshot(header.Number)
+
 	if snapshot == nil || metaEpoch < currentEpoch {
 		// Restore snapshot at the beginning of the current epoch by block header
 		// if list doesn't have any snapshots to calculate snapshot for the next header
 		i.logger.Info("snapshot was not found, restore snapshot at beginning of current epoch", "current epoch", currentEpoch)
 		beginHeight := currentEpoch * i.epochSize
 		beginHeader, ok := i.blockchain.GetHeaderByNumber(beginHeight)
+
 		if !ok {
 			return fmt.Errorf("header at %d not found", beginHeight)
 		}
@@ -61,6 +64,7 @@ func (i *Ibft) setupSnapshot() error {
 		if err := i.addHeaderSnap(beginHeader); err != nil {
 			return err
 		}
+
 		i.store.updateLastBlock(beginHeight)
 
 		if meta, err = i.getSnapshotMetadata(); err != nil {
@@ -76,10 +80,12 @@ func (i *Ibft) setupSnapshot() error {
 			if num == 0 {
 				continue
 			}
+
 			header, ok := i.blockchain.GetHeaderByNumber(num)
 			if !ok {
 				return fmt.Errorf("header %d not found", num)
 			}
+
 			if err := i.processHeaders([]*types.Header{header}); err != nil {
 				return err
 			}
@@ -138,6 +144,7 @@ func (i *Ibft) processHeaders(headers []*types.Header) error {
 	if err != nil {
 		return err
 	}
+
 	snap := parentSnap.Copy()
 
 	// saveSnap is a callback function to set height and hash in current snapshot with given header
@@ -230,6 +237,7 @@ func (v *Vote) Equal(vv *Vote) bool {
 func (v *Vote) Copy() *Vote {
 	vv := new(Vote)
 	*vv = *v
+
 	return vv
 }
 
@@ -261,6 +269,7 @@ func (s *Snapshot) Equal(ss *Snapshot) bool {
 	if len(s.Votes) != len(ss.Votes) {
 		return false
 	}
+
 	for indx := range s.Votes {
 		if !s.Votes[indx].Equal(ss.Votes[indx]) {
 			return false
@@ -278,6 +287,7 @@ func (s *Snapshot) Count(h func(v *Vote) bool) (count int) {
 			count++
 		}
 	}
+
 	return
 }
 
@@ -313,7 +323,7 @@ func (s *Snapshot) ToProto() *proto.Snapshot {
 	resp := &proto.Snapshot{
 		Validators: []*proto.Snapshot_Validator{},
 		Votes:      []*proto.Snapshot_Vote{},
-		Number:     uint64(s.Number),
+		Number:     s.Number,
 		Hash:       s.Hash,
 	}
 
@@ -362,6 +372,7 @@ func (s *snapshotStore) loadFromPath(path string) error {
 	if err := readDataStore(filepath.Join(path, "metadata"), &meta); err != nil {
 		return err
 	}
+
 	if meta != nil {
 		s.lastNumber = meta.LastBlock
 	}
@@ -371,6 +382,7 @@ func (s *snapshotStore) loadFromPath(path string) error {
 	if err := readDataStore(filepath.Join(path, "snapshots"), &snaps); err != nil {
 		return err
 	}
+
 	for _, snap := range snaps {
 		s.add(snap)
 	}
@@ -467,6 +479,7 @@ func (s *snapshotStore) replace(snap *Snapshot) {
 	for i, sn := range s.list {
 		if sn.Number == snap.Number {
 			s.list[i] = snap
+
 			return
 		}
 	}
@@ -515,6 +528,7 @@ func writeDataStore(path string, obj interface{}) error {
 		return err
 	}
 
+	//nolint: gosec
 	if err := ioutil.WriteFile(path, data, 0755); err != nil {
 		return err
 	}
